social_media_api/requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.1
django-crispy-forms==2.3
django-filter==24.3
django-taggit==6.0.0
django_csp==3.8
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
